pyproject.toml
[project]
name = "multi-agent-conversational-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
        "agentops>=0.4.21",
    "autogen>=0.9.9",
    "chromadb>=1.1.0",
    "deepeval>=3.3.9",
    "dspy>=3.0.3",
    "fastapi>=0.118.0",
    "gradio>=5.49.0",
    "gradio-tools>=0.0.9",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "langchain[openai]>=1.0",
    "langchain-community>=0.3.30",
    "langchain-openai>=0.3.34",
    "langgraph>=0.6.8",
    "matplotlib>=3.10.6",
    "openai>=2.0.0",
    "openai-gradio>=0.0.4",
    "py-trees>=2.3.0",
    "pydantic>=2.11.9",
    "python-dotenv",
    "scikit-learn>=1.7.2",
    "semantic-kernel>=1.36.0",
    "streamlit>=1.50.0",
    "unstructured>=0.18.15",
    "uvicorn>=0.37.0",
    "requests>=2.32.5",
    "jq",
    "sentence-transformers>=5.1.1",
    "tqdm>=4.67.1",
    "hf-xet>=1.1.10",
    "adjusttext>=1.3.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.43",
    "psycopg2>=2.9.11",
    "fastmcp>=2.12.5",
    "pypdf>=6.1.1",
    "numexpr>=2.14.1",
    "langchain-tavily>=0.2.12",
    "ngrok>=1.4.0",
    "langchain-mcp-adapters>=0.1.12",
    "langgraph-api>=0.4.48",
    "langsmith>=0.4.31",
    "langchain-text-splitters>=1.1.0",
]
